Permisime te vogla neper kontrollera dhe shfaqja e pending invites te realizuara ne nje workspace
- Jane bere disa permisime neper disa kontrollera per te larguar problemet e null possibility reference
- eshte shtuar nje pjese ne faqen members e cila shfaqet vetem per Owner-in e workspaces ne fjale ku listohen te gjitha pending invites qe jane realizuar ne ate workspace dhe gjithashtu te dhenat e invitee
- si dhe per secilen pending invite nje button delete qe owneri te mund te fshij ate invite komplet

- funksionaliteti per ti marr te dhenat e pending invite eshte bartur ne workspace context qe te mund te shfrytezohet qe ne momentin kur te behet nje invite i ri te shfaqet menjeher ne ate liste

+++ backend/Controllers/MembersController.cs
@@ -46,6 +46,11 @@ public class MembersController: ControllerBase
             }
             var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+            
+            if (string.IsNullOrEmpty(userId))
+            {
+                return NotFound("User Not Found!");
+            }
             var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, addMemberDto.WorkspaceId);
             if (ownsWorkspace || userTokenRole == "Admin")
             {
@@ -95,6 +100,11 @@ public class MembersController: ControllerBase
             }
             var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+            
+            if (string.IsNullOrEmpty(userId))
+            {
+                return NotFound("User Not Found!");
+            }
             var isMember = await _membersRepo.IsAMember(userId, workspaceId);
             if (isMember || userTokenRole == "Admin")
             {
@@ -137,15 +147,19 @@ public class MembersController: ControllerBase
             }
             var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+            
+            if (string.IsNullOrEmpty(userId))
+            {
+                return NotFound("User Not Found!");
+            }
              var isMember = await _membersRepo.IsAMember(userId, removeMemberDto.WorkspaceId);
             if (isMember || userTokenRole == "Admin")
             {
                 var result = await _membersRepo.RemoveMemberAsync(removeMemberDto.WorkspaceId, removeMemberDto.UserId);
                 if (result == null)
                 {
-                    return StatusCode(500, "User could not removed");
+                    return StatusCode(500, "User could not be removed");
                 }
-
                 return Ok("Member removed!");
             }
             return StatusCode(401, "You are not authorized!");
@@ -225,6 +239,10 @@ public class MembersController: ControllerBase
         {
             var userId = User.Claims.FirstOrDefault(c => c.Type == "Id")?.Value;
             var userTokenRole = User.Claims.FirstOrDefault(c => c.Type == "Role")?.Value;
+            if (string.IsNullOrEmpty(userId))
+            {
+                return NotFound("User Not Found!");
+            }
             var ownsWorkspace = await _userRepo.UserOwnsWorkspace(userId, workspaceIdDto.WorkspaceId);
             if (ownsWorkspace || userTokenRole == "Admin")
             {
